requirements.txt
flask
flask-cors
websockets
python-dotenv
openai
PyMuPDF
